move enable_zluda_config into require_torch
